Decode instruction 0xANNN
For now, setting the type of registers to int.  Even a machine as simple
as a CHIP8 can have more than one width of register, so the previous
global def approach won't work;  not sure how to handle right now, so in
the spirit of TDD I'm making a note and moving on.

I'm also not wild about the way that GTest seems to do parameterized tests
(pytest has a much cleaner implementation), but this may just be the
typical C++ verbosity.