Decode instruction 0x7XNN
A utility to generate opcodes of form XYNN has been added to reduce
copy-and-paste when testing instructions that target all registers,
and the previous 0x6XNN tests have been ported over to using it.

This method of auto-generating tests introduces a for loop into the
body of the test, which will make debugging more difficult down the
line, but the reduction in verbosity is worth it.