Remove dummy singleton invocation
After doing some inspection, the crash when running
RetroArch revolved around retro_get_memory_size returning a
positive value.  Looking through the source code, the main
function I could find for this variable is the autosave
functionality, in particular skipping it if the reported
memory size is zero or negative (as I've done here.)  This
suggests that the autosave and serialization functionality
may be the culprit.  Why the dummy singleton prevents this, I
don't know.

In addition, I've also implemented the retro_get_memory_data()
subroutine.  This was a dead-in in debugging, but I figure I
might as well keep it in.